ci: adds reusable setup polytest and mock server actions, minor scripts and docs refinements (#18)
* chore: cd for image publishing

* chore: test (#1)

* fix: handles binary msgpack responses correctly

Ensures correct handling of binary msgpack responses by using `arrayBuffer()` instead of `text()` to prevent UTF-8 corruption.
Also adds logging for content type and binary status, and adjusts the response preview logic to only apply to non msgpack data.

* chore: cd for image publishing

* feat: add shared setup-polytest gh action

* chore: Update test_configs/kmd_client.jsonc

* chore: update .github/workflows/publish-docker.yml

* refactor: adds run-mock-server GitHub Action

Introduces a composite GitHub Action to simplify starting mock servers for testing Algorand API clients (algod, indexer, kmd).

This action handles installing Bun, mock-server dependencies,
starting the server in the background, waiting for the server to be healthy, and exporting environment variables for use in subsequent steps.

It uses the mock-server located in the `resources/mock-server/` directory of this repo and makes it available as a reusable action.

* chore: tweak health check

* docs: tweak docs

* chore: copilot comments